echoscript-backend-src-unzipped/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Core typing/config
pydantic==2.9.2
pydantic-settings==2.6.0

# Web & uploads
python-multipart==0.0.9
lxml==5.2.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# DB / migrations (psycopg2-binary is fine on Railway)
sqlalchemy==2.0.34
alembic==1.13.2
psycopg2-binary==2.9.9

# Payments
stripe==11.*

# Utilities
redis==5.0.8
httpx==0.27.2
loguru==0.7.2

# Export / media
python-docx==0.8.11
fpdf==1.7.2
ffmpeg-python==0.2.0

# --- Transcription dependencies (enabled) ---
faster-whisper==1.0.3
ctranslate2==4.5.0
sentencepiece==0.2.0
tokenizers==0.19.1
numpy==1.26.4
email-validator==2.2.0

